Added in optional removal of URLs to txts2txt.py
Added in optional removal of URLs to txts2txt.py to allow removal of whole URLS (seems to work automagickly).
TODO:
Need to migrate filter into function to make it easier to use on all the texts.
Also need to allow modified filter to allow just the url name to remain.